SI-4859 Don't elide qualifiers when selecting nested modules.
Otherwise we fail to throw in:

   {???; Predef}.DummyImplicit.dummyImplicit

We still elide the initialization of `Outer` in `Outer.Inner.foo`
as before, although that seems a little dubious to me.

In total, we had to change RefChecks, Flatten, and GenICode
to effect this change. A recently fixed bug in tail call elimination
was also due to assuming that the the qualifier of a Select node
wasn't worthy of traversal.  Let's keep a close eye out for more
instances of this problem.

Furthermore, dead code elimination eliminates the module load
of `Outer` in `Outer.Inner.foo`, which would cause run/t4859.scala
to fail under the -optimize. We now err on the side of correctness and retain all MODULE_LOAD instructions unless a backwards
compatibility flag is set.