Remove name-mangling for almost all characters
While the list of exceptions hasn't changed much in size,
non-dangerous method names like + and ??? as well as 99% of the
character plane which were formerly name-mangled with $uXXXX appear
now as-is in stacktraces, other places and tools not specificially
written with Scala's name-mangling in mind.

The changes are mostly in line with the recommendations for the JVM
(look for "symbolic freedom"), except for the list of exceptions
which currently keep using the $-style mangling instead of the suggested
backslash-style (example: / is currently encoded as $slash, the
guidelines recommend \|) due to readability concerns.

The improvements to readability are impressive, imho.
Another benefit is that people have to deal with a lot less mangled
names when working with macros.