Fixes and features for javap (fixing SI-6894)
Output filtering is refactored for javap6.

That means javap6 also supports -raw.

Handling of # is: Foo#foo filter on foo,
Foo# filter on apply, -fun Foo#foo for anonfuns of foo,
-fun Foo# anonfuns filtering on apply.

One is loath to add command options, so it's not possible
to ask for "only apply methods in anonfuns pertaining to
a method."  Hypothetical syntax to say "show me the
apply only": -fun Foo#foo(), for future reference.