SI-6987 Fixes race condition in testing
On the build server the test failed once but passed other times. There's
a race condition between server start up and client start up. The fsc
client normally deals with that via a retry/timeout mechanism but in
testing environment with lots of threads and processes going on that
wasn't sufficient. This commit adds a server callback that decrements a
CountDownLatch that the main testing thread waits for.

Along the way I removed an unnecessary redundant catch block from the
test.