Always build scala-swing in java7 branch.
The 6dd90d349da6471c0a105c584798979f136d7d55 commit introduced a way
to skip building swing in case one is not using Java 6 (e.g. using
Java 7). That change was meant to make it easier to work with master
no matter if one is using Java 6 or Java 7 because for everything
apart from swing it doesn't matter.

However, that change propagated to java7 branch which exists only to
deal with Swing problems and is meant to be used with Java 7 only.
In this branch we should get rid of conditional building of Swing and
build all the time.