SI-6728 Fixes crash in parser on incomplete for expression
The parser was trying to take the position of an empty expression
which would crash. Son on the empty expression case in TreeBuilder#
makeWhile it tries to do that and, if that failed, gets a position
that wraps both the condition and the body.

I also made a slight improvement to the UnsupportedOperationEx
messages in Position.