SI-3353 don't extract <unapply-selector> into named-arg local val
This already fixes all of SI-3353 because no named-args-block is
generated anymore (they are avoided if they have a single expr).

So the same NPE in extractorFormalTypes as described in the ticket
is no longer triggered.

I think that's all there is to fix, since extractor patterns are
translated to unapply calls with one argument, i think it's not
possible to write a pattern that would result in a named-apply block.